+++ CHANGELOG.md
@@ -14,7 +14,7 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 #### Minor Changes
 
 - Extend the Wrapper API
-  - Adds `createDataOffer` endpoint to create an asset, policies and a contract definition ina single call
+  - Adds `createDataOffer` endpoint to create an asset, policies and a contract definition in a single call
 
 #### Patch Changes