+++ CHANGELOG.md
@@ -20,12 +20,32 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Patch
 
-- Changed wording on the data offer creation page
-- Data Offer details now display the contract ID for each contract offer
-- Added validator for the Policy Id
-- Added validator for the Data Offer Id
+### Deployment Migration Notes
+
+_No special deployment migration steps required_
+
+## [v4.1.4] - 2024-09-17
+
+### Overview
+
+MDS Patch release
+
+### Detailed Changes
+
+Various bugfixes.
+
+#### Patch
+
+- Changed wording on the data offer creation page ([#817](https://github.com/sovity/edc-ui/issues/795))
+- Data Offer details now display the contract ID for each contract offer ([#795](https://github.com/sovity/edc-ui/issues/795))
+- Warn the user when using an invalid Policy Id ([#746](https://github.com/sovity/edc-ui/issues/746))
+- Warn the user when using an invalid Data Offer Id ([#745](https://github.com/sovity/edc-ui/issues/745))
 - Fixed time restriction upper bound "local day to datetime" conversion issues
   ([#815](https://github.com/sovity/edc-ui/issues/815))
+
+### Deployment Migration Notes
+
+_No special deployment migration steps required_
 
 ## [v4.1.3] - 2024-09-03